Chat modify1

+++ BackEnd/src/main/java/com/se/EC/Service/Chat/ChatService.java
@@ -3,10 +3,13 @@ package com.se.EC.Service.Chat;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.github.jeffreyning.mybatisplus.service.MppServiceImpl;
 import com.se.EC.Controller.Chat.Information;
+import com.se.EC.Controller.Chat.SessionInformation;
 import com.se.EC.Entity.Chat;
 import com.se.EC.Entity.Session;
+import com.se.EC.Entity.User;
 import com.se.EC.Mapper.ChatMapper;
 import com.se.EC.Mapper.SessionMapper;
+import com.se.EC.Mapper.UserMapper;
 import jakarta.annotation.Resource;
 import org.springframework.stereotype.Service;
 
@@ -21,6 +24,8 @@ public class ChatService extends MppServiceImpl<ChatMapper, Chat> implements Cha
     private ChatMapper chatMapper;
     @Resource
     private SessionMapper sessionMapper;
+    @Resource
+    private UserMapper userMapper;
 
     /**
      * 创建会话
@@ -32,10 +37,14 @@ public class ChatService extends MppServiceImpl<ChatMapper, Chat> implements Cha
     public void createSession(Integer senderId, Integer receiverId) {
         // 查看这个会话是否已经存在
         Long count = sessionCount(senderId, receiverId);
+        Long count1 = sessionCount(receiverId, senderId);
 
-        if (count == 0) {  // 如果不存在就新建Session，并返回True
-            Session session = new Session(senderId, receiverId, LocalDateTime.now());
+        if (count == 0 && count1 == 0) {  // 如果不存在就新建Session，并返回True
+            LocalDateTime currentTime = LocalDateTime.now();
+            Session session = new Session(senderId, receiverId, currentTime);
+            Session session1 = new Session(receiverId, senderId, currentTime);
             sessionMapper.insert(session);
+            sessionMapper.insert(session1);
         } else if (count != 1) {  // 多于一个出错
             throw new RuntimeException("There are two session between " + senderId + " and " + receiverId);
         }
@@ -55,13 +64,45 @@ public class ChatService extends MppServiceImpl<ChatMapper, Chat> implements Cha
         queryWrapper.eq("chat_receiver", receiverId);
         Long count = sessionMapper.selectCount(queryWrapper);
 
-        if (count == 0) {  // 如果不存在返回错误
+        QueryWrapper<Session> queryWrapper1 = new QueryWrapper<>();
+        queryWrapper.eq("chat_receiver", senderId);
+        queryWrapper.eq("chat_sender", receiverId);
+        Long count1 = sessionMapper.selectCount(queryWrapper);
+
+        if (count == 0 && count1 == 0) {  // 如果不存在返回错误
             throw new RuntimeException("Session does not exist");
-        } else if (count == 1) {  // 如果已经存在就删除
+        } else if (count == 1 && count1 == 1) {  // 如果已经存在就删除
             sessionMapper.delete(queryWrapper);
+            sessionMapper.delete(queryWrapper1);
         } else {  // 如果多于一个返回错误
             throw new RuntimeException("There are many session between " + senderId + " and " + receiverId);
         }
+    }
+
+    /**
+     * 获取会话
+     *
+     * @param id id
+     */
+    @Override
+    public List<SessionInformation> getSession(Integer id) {
+        // 获取所有会话
+        QueryWrapper<Session> queryWrapper = new QueryWrapper<>();
+        queryWrapper.eq("chat_sender", id);
+        List<Session> sessionList = sessionMapper.selectList(queryWrapper);
+
+        // 转换
+        List<SessionInformation> sessionInformationList = new ArrayList<>();
+        for (var item : sessionList) {
+            Integer receiverId = item.getReceiverId();
+            QueryWrapper<User> userQueryWrapper = new QueryWrapper<>();
+            queryWrapper.eq("user_id", receiverId);
+            User user = userMapper.selectOne(userQueryWrapper);
+            SessionInformation sessionInformation = new SessionInformation(receiverId, user.getUser_name());
+            sessionInformationList.add(sessionInformation);
+        }
+
+        return sessionInformationList;
     }
 
     /**
